More flexible indexing into Array1D
- users can set up indexing for `Range`s, i.e. indexng does not have
  to start from 0
